// packages/ui-mini/src/InputNumber/InputNumber.types.ts
export type InputNumberSize = 'sm' | 'md' | 'lg'

export interface InputNumberProps {
  value?: number | null
  defaultValue?: number | null
  onChange?: (value: number | null) => void
  min?: number
  max?: number
  step?: number
  precision?: number
  disabled?: boolean
  size?: InputNumberSize
  placeholder?: string
  controls?: boolean
  className?: string
}